Move LHCb2 style to new location
