+++ src/mplhep/lhcb2.py
@@ -1,6 +1,5 @@
 # Log styles
-from . import styles_lhcb2 as style
+from mplhep.styles import lhcb2 as style
 from .label import lumitext
-from .lhcb_common import label, text
 
 __all__ = [style, lumitext]
