+++ src/mplhep/styles/lhcb2.py
